requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.9.3
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.4.1
pytest-django==3.4.8
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
yapf==0.27.0
